[feat] using allo matmul lib for gemm on XDNA1
